Refactor and adding tracing implementation - Add dependencies for tracing implementation - Refactor code via building jar and running it, during running found couple of errors and fixing them - Bootify and beautify code
