+++ app/kafka/src/main/java/com/simple/kafka/publisher/SimpleKafkaPublisherImpl.java
@@ -2,14 +2,8 @@ package com.simple.kafka.publisher;
 
 import static com.simple.kafka.config.SimpleKafkaConfig.PAYLOAD_TYPE_HEADER;
 
-import java.util.Objects;
-import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.support.KafkaHeaders;
-import org.springframework.kafka.support.SendResult;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.stereotype.Component;
@@ -20,7 +14,6 @@ import com.simple.service.api.converter.SimpleKafkaConverter;
 import com.simple.service.api.publisher.SimpleKafkaPublisher;
 
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
