+++ app/rest/build.gradle
@@ -5,4 +5,6 @@ dependencies {
     implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:${springdoc_openapi_version}"
     implementation "org.springframework.boot:spring-boot-starter-web"
     implementation "jakarta.validation:jakarta.validation-api:${validation_jakarta_version}"
+
+    implementation "io.micrometer:micrometer-registry-prometheus:${metrics_prometheus_version}"
 }
