+++ app/rest/src/main/java/com/simple/rest/config/SimpleControllerConfig.java
@@ -2,8 +2,10 @@ package com.simple.rest.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @Configuration
 @ComponentScan(basePackages = "com.simple.rest")
+@EnableWebMvc
 public class SimpleControllerConfig {
 }
