+++ app/service-api/src/main/java/com/simple/service/api/date/SimpleDateProviderService.java
@@ -5,6 +5,5 @@ import java.time.ZonedDateTime;
 
 public interface SimpleDateProviderService {
     DateFormat getDateFormat();
-    ZonedDateTime currentDate();
-
+    ZonedDateTime now();
 }
