+++ app/service/src/main/java/com/simple/service/config/SimpleServiceConfig.java
@@ -1,9 +1,23 @@
 package com.simple.service.config;
 
+import java.text.SimpleDateFormat;
+
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
 @Configuration
 @ComponentScan(basePackages = "com.simple.service")
 public class SimpleServiceConfig {
+    @Bean
+    public ObjectMapper objectMapper() {
+        final ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.registerModule(new JavaTimeModule());
+        final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
+        objectMapper.setDateFormat(dateFormat);
+        return objectMapper;
+    }
 }
