--- app/src/main/java/com/simple/main/config/SimpleApplicationConfig.java
@@ -1,28 +1,0 @@
-package com.simple.main.config;
-
-import java.text.SimpleDateFormat;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.simple.kafka.config.SimpleKafkaConfig;
-import com.simple.service.api.date.SimpleDateProviderService;
-
-@Configuration
-@EnableWebMvc
-@Import(SimpleKafkaConfig.class)
-public class SimpleApplicationConfig {
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        final ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.registerModule(new JavaTimeModule());
-        final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
-        objectMapper.setDateFormat(dateFormat);
-        return objectMapper;
-    }
-}
