+++ app/src/main/resources/application.properties
@@ -9,7 +9,7 @@ spring.data.mongodb.port=27017
 spring.data.mongodb.auto-index-creation=true
 
 spring.mvc.throw-exception-if-no-handler-found=true
-management.endpoints.web.exposure.include=*
+management.endpoints.web.exposure.include=prometheus
 management.endpoints.web.cors.allowed-origins=http://localhost:8080
 management.endpoints.web.cors.allowed-methods=*
 
@@ -36,3 +36,7 @@ spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.B
 spring.kafka.producer.bootstrap-servers=${spring.kafka.bootstrap-servers}
 spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.ByteArraySerializer
+
+# OpenTelemetry
+spring.sleuth.otel.config.trace-id-ratio-based=1.0
+spring.sleuth.otel.exporter.otlp.endpoint=http://collector:4317
