+++ build.gradle
@@ -1,7 +1,9 @@
+apply from: "${project.rootDir}/gradle/config/service.gradle"
+
 buildscript {
 	repositories {
+		mavenCentral()
 		maven {
-			mavenCentral()
 			url "https://plugins.gradle.org/m2/"
 		}
 	}
@@ -11,5 +13,3 @@ buildscript {
 		classpath "com.bmuschko:gradle-docker-plugin:${gradle_docker_plugin_version}"
 	}
 }
-
-apply from: "${project.rootDir}/gradle/config/service.gradle"
