+++ docker/Dockerfile
@@ -1,0 +1,5 @@
+FROM eclipse-temurin:17
+ARG JAR_FILE=target/*.jar
+COPY ${JAR_FILE} application.jar
+EXPOSE 8080
+ENTRYPOINT ["java","-jar","/application.jar"]
