+++ docker/otel-config.yml
@@ -1,0 +1,23 @@
+receivers:
+  prometheus:
+    config:
+      scrape_configs:
+        - job_name: "example"
+          scrape_interval: 5s
+          metrics_path: '/actuator/prometheus'
+          static_configs:
+            - targets: ["localhost:8080"]
+
+processors:
+  batch:
+
+exporters:
+  prometheus:
+    endpoint: "localhost:8889"
+
+service:
+  pipelines:
+    metrics:
+      receivers: [prometheus]
+      processors: [batch]
+      exporters: [prometheus]
