+++ gradle.properties
@@ -1,15 +1,33 @@
+
 mapstruct_version=1.5.5.Final
+
 lombok_mapstruct_binding_version=0.2.0
-jackson_version=2.15.1
-springframework_version=3.1.1
-gradle_docker_plugin_version=6.7.0
-junit_bom_version=5.9.1
+
 slf4j_version=2.0.7
-mockito_version=5.4.0
+
+jackson_version=2.15.1
+
 jackson_datatype_version=2.15.1
-validation_jakarta_version=3.0.2
+
+springframework_version=3.1.1
+
 spring_dependecy_management_version=1.1.0
+
+gradle_docker_plugin_version=6.7.0
+
+junit_bom_version=5.9.1
+
+mockito_version=5.4.0
+
+validation_jakarta_version=3.0.2
+
 test_containers_version=1.18.3
 
 # [SpringDoc]
 springdoc_openapi_version=2.1.0
+
+# [OpenTelemetry]
+opentelemetry_version=1.19.0
+
+# [Prometheus - grafana]
+metrics_prometheus_version=1.11.2
